requirements.txt
pandas==2.0.1
python-dotenv==1.0.0
Requests==2.31.0
selenium==4.9.1
